Write commit graph with split strategy
Re-writing of the full commit graph is an expensive
operation. We are going to use a split strategy with
size of 4 (https://git-scm.com/docs/git-commit-graph#Documentation/git-commit-graph.txt-emwriteem)
to reduce time required to change a commit graph.
The configuration for gc operation changed from mutating
persistent config with per-operation configuration for gc.