Remove read only repository metrics collector
Scrapping metrics from the database should be extracted
into a separate deployment as it prevent horizontal scaling
of the praefect instances. As part of praefect it could
produce too much pressure on the database if we begin to scale
it out. It also will produce too many metrics and create
additional pressure on the collector. The solution for that
problem is usage of additional exporter that will be deployed
independently from the praefect instance.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3286