datastore: Clean completed & dead replication jobs
In [1] changes were made so 'completed' and 'dead' jobs are no longer
kept in the database, and are cleaned up. In a follow-up commit [2] a
migration was added to clean up all 'completed' and 'dead' jobs from the
replication queue. But this method did not take the order of deployment
into account. So the migration did run before the code was deployed,
which resulted some new 'completed' and 'dead' rows to be added for a
little while.

This change repeats the `DELETE` query from the migration added in [2]
to clean up 'completed' and 'dead' rows once and for all.

1. 8f8ae3025 (Replication job acknowledge removes 'completed' and 'dead'
   events., 2020-08-05)
2. 2d3fc8063 (Replication job acknowledge removes 'completed' and 'dead'
   events., 2020-09-03)

Changelog: performance
Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3665