Praefect uses listen_addresses as list of addresses to listen on
To simplify and reduce amount of configuration settings
the new listen_addresses configuration value is added.
The intent is to replace existing socket_path, listen_addr and
tls_listen_addr configurations with a single one.
It also allows to use more then one address for a single schema.
The existing configurations fields marked as deprecated and
would be removed once the migration process is done.
In case both the old and the new configuration values are set
the old values will be used and new values will be omitted.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2927