Run Postgres as a Docker container from the code
For a long period we were using in-memory implementation
of the job queue because spinning up a container and managing
of it seemed to be an overhead for the unit-tests.
But with the testcontainers package it seems to be an easy to
do. The change adds a RunPostgres function that spins up a
Docker container with Postgres database and runs migration on
top of it. It doesn't conflict with other running containers
as it picks an exposed port address automatically.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2636