memory: Pool of bytes buffers
In order to reduce memory allocations and better utilise
already allocated memory it makes sense to use a shared
pool of *bytes.Buffer instances. The solution uses a standard
sync.Pool that is used in stdlib for the same purpose.
The Buffer() function acquires a new or existing instance
from the pool and a release function that resets content
of the buffer and returns it back to the pool.