memory: Use pool of *bytes.Buffer
In order to reduce memory allocations and better utilise
already allocated memory we now start to use a shared
pool of *bytes.Buffer instances. We replace the local
instantiation of the *bytes.Buffer with call to helper.Bytes()
that returns a new or existing instance. The replacement
done in case the buffer is not used once the function returns.
In case the data needs to be preserved we can't use shared
pool as after call to release function the data would be
abandoned. The usage of the pre-existing buffer should reduce
memory allocation and potentially speed up execution as well
as reduce GC pauses and invocations.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3791

Changelog: performance