migrations: Ignore repository_id in UPDATE trigger
The `20210906145021_link_repository_id` migration touches all rows in
`storage_repositories`. This causes the TRIGGER ON UPDATE on that table
to pass a very large payload to pg_notify when there a lot of
repositories known to Praefect.

This change modifies that TRIGGER to exclude changes that only touch the
`repository_id`.

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3806