migrations: Revert backfilling repository_id
In [1] a migration was added to backfill the repository_id on the
storage_repositories table. That backfill has been using a single UPDATE
query to update all records at once. But because there is a TRIGGER on
that table it caused a PG_NOTIFY to be sent with a payload that exceeds
the maximum allowed by PostgreSQL, and the migration to fail on large
datasets [2].

That commit is part of [3]. After that commit, the queries can be
modified to JOIN tables on the repository_id, instead of on the
virtual_storage + relative_path. But those changes are not merged
yet. Therefore this change removes the migration to unblock
installations with large datasets, until a proper solution to backfill
the records is found.

1. 2bbec66c9 (Link existing database record via repository ID,
   2021-09-01)
2. https://gitlab.com/gitlab-org/gitaly/-/issues/3806
3. https://gitlab.com/gitlab-org/gitaly/-/issues/3485