replication: Remove 'dead' stale jobs.
When replication job processing is completed we do
acknowledgment of the event. If processing was successful
the state of the event should be changed to 'completed',
for failed processing attempts the state becomes 'failed'
and if event processing failed max amount of attempts it
should be changed to 'dead'. To reduce amount of rows in
the replication_queue table the events that are going to
be moved into 'completed' or 'dead' state not updated but
removed from the table. The change aligns this behavior to
AcknowledgeStale method that was not remove 'dead' events
from the table, but only update their status.