streamio: Track amount of processed bytes
The streamio package is used to work with streaming
gRPC methods. It provides a wrappers that allow to
receive and send data from/to gRPC channel if it is
io.Reader or io.Writer. The change adds a new WrittenBytes
and ReadBytes structs that wraps standard io.Reader and
io.Writer interfaces and calculates amount of transferred
bytes. The Total method returns amount of bytes processed
by the wrappers.