sql: Drop 'cancelled' state from REPLICATION_JOB_STATE
The state 'cancelled' is not used in the production
code and no any queries are using it as well. The
database should be aligned with Go code and unused
value should be dropped from the enum. As the enum
type is used in index and view those entities need
to be re-created with the new representation of the
enum type.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3859