testhelper: Make gitaly-git2go symlink race-free
The compilation of binaries was made race-free in 16d52986c (testhelper:
Refactor building of binaries for improved readability, 2021-10-08), but
the creation of the symlink to the module-versioned path still suffered
from races.

This change makes the symlink part of the compilation step, guarded with
a sync.Once.

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3869