ci: Use non-merged result SHA for QA pipeline
When the merged result pipeline is triggered it clones Gitaly. But it
does not fetch the merged result revision. Ideally we should change the
downstream pipeline so it fetches that revision, but for now just
checkout the HEAD revision on the merge request source branch.

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3852