ci: Pass repository URL to QA pipeline
A child QA pipeline is triggered with ALTERNATIVE_SOURCES=true and a
GITALY_SERVER_VERSION set to the HEAD of the source branch of the merge
request. This makes QA expect the revision to be on
gitlab.com/gitlab-org/gitaly, but in some cases it's not
guaranteed (i.e. changes made to the security-release fork).

This change passes on GITALY_SERVER_ALTERNATIVE_REPO so the QA project
pulls Gitaly from where the pipeline was triggered and will find the
revision specified.