ci: Run QA always on the default branch
To have a better understanding of how the changes behave in
gitlab-org/gitlab, this change triggers to QA pipeline for all changes
on the default branch.

In case merge trains are enabled, the merged result is not on master
yet, so also here GITALY_SERVER_VERSION is set to the HEAD commit of the
source branch.