datastore: Split out functionality to get Postgres version
We'll need the ability to check the Postgres version at runtime so that
we can decide whether it supports `MATERIALIZED` or not. Split out the
logic to detect the version from `CheckPostgresVersion()` into its own
function `getPostgresVersion()` to make this feasible.

Note that `getPostgresVersion()` uses a `sync.Once` to retrieve the
version once only: the database version we use is not expected to change
at runtime, and if it does then the expectation would be that it is
never a downgrade. Given that this check is going to be used to verify
that Postgres supports some new features, the worst that could happen is
that we not use those new features.

Further note that this isn't really well-testable. Ideally, this would
be part of a Gitaly-specific database structure, but outside of our
tests we don't have that. Given that the fix that will make use of this
function should land comparatively quickly to fix a regression, these
wider refactorings are left for a later point.