glsql: Apply migrations partially
To verify change done by particular migration we
need to apply only all previous migrations first.
The test database if requested with not all migrations
won't use a template database, but will create a new
one and apply only a subset of all migrations.
Once setup is done the remaining migrations could be
applied to check if they do what is expected.