ci: Run jobs as unprivileged user
Our CI jobs are currently running as the privileged "root" user. This
user is special in many different ways, but most importantly it can use
a bunch of capabilities by default. Most importantly in our context, the
user has the DAC_OVERRIDE capability by default, which will allow it to
disregard file permission bits completely and just do read or write
files even though a normal user wouldn't be able to. This is of course
different to a normal Gitaly installation and to developer setups, where
Gitaly really shouldn't ever run with special privileges. As a result,
it has caused us to not detect some issues with permissions in our tests
which would've been catched if we ran jobs as unprivileged user.

Convert our CI jobs to instead create and use an unprivileged user. Note
that this requires us to use a new prefix for all of our caches given
that the old caches would've been created by the root user so that their
contents aren't writable by the current user.