Add size in GitEntry
The GetTreeEntries method in commit service is used to retrieve git entries information.
Currently, depending on the recursive params, git-ls-tree or git-cat-file is used.
This PR MR enables the -l param in git-ls-tree in order to extract the Entry size.
I've check in git-cat-file, but seems like the Tree object, when displayed, only report Oid, Mode and Filename
So this PR enable the use of git-ls-tree for listing and git-cat-file for flatpath compute.