cmd/praefect: Display full replication status by default in 'dataloss' subcommand
Because dataloss command is used to show the full state of repository replication
is more common than only showing the repository with an unhealthy primary. So
make the default command behavior  show all repositories with full unavailable
and partially unavailable replication status.

This commit keep the dataloss command's partially-unavailable flag available  for
backward compatibility. And introduce the fully-unavailable flag for only showing
the repositories with an unhealthy primary.