Add stat option for findcommits to get stats of commit
To calculate the contribution information of each contributor, currently we need to call `FindCommits` to get the latest commits, and call `CommitStats` multiple times to get `adds/dels`.

The problem is: it takes too much time when querying `adds/dels` for a large number of commits.

To solve the N + 1 problem, we propose to add a flag `--shortstat` for `git log` command in `FindCommits`, so that we do not need to call `git diff` for each commit.

Add stat for changed files.