git: Introduce feature flag for hiding refs in git-upload-pack(1)
We have observed an unexpected CPU spike as well as an increase in write
operations after 4a789524c (git: Don't advertise internal references via
git-upload-pack(1), 2022-07-19) has been deployed, which causes us to
stop advertising some internal references to clients which fetch from a
repository. It's unexpected that this change increases such a spike, and
thus we're being careful for now and introduce a feature flag so that we
can ease into the rollout of this change while being able to roll it
back with ease.