Return context error from requestFinalizer
Praefect runs the request finalizer even if the request deadline was
exceeded in order to ensure the required metadata updates are made.
If the context deadline is exceeded in the request finalizer, no error
is returned as the metadata updates don't see it. This causes Praefect
to return, log and observe a status OK for requests that exceeded the
deadline or were canceled while running the request finalizer. This
commit returns the error from the original context if no other errors
are raised from the requestFinalizer to ensure the status gets correctly
set.

desc field is added to the tests as they currently abuse the change field
as the test name.

Changelog: fixed