gitlab: Copy `client` package from `gitlab-shell`
To remove the dependency on `gitlab-shell` we want to internalize the
`client` package. While the ideal step is to move this to the `labkit`
repository, we're taking an intermediate approach wherein we're copying
this code inhouse to our repo and mantaining it in parallel.

Note: This commit directly copies the package 1:1, without
modifications by running `cp -r
$GOPATH/pkg/mod/gitlab.com/gitlab-org/gitlab-shell/v14@v14.10.0/client/
internal/gitlab/client/`. This is breaking commit because some of the
tests depend on internal code of `gitlab-shell`, the follow up commits
will be to fix/cleanup those tests and remove any external dependency.

Signed-off-by: Karthik Nayak <knayak@gitlab.com>