Makefile: Fix target that lints our linter
The shared object for our Gitaly-specific linter is implemented in its
own module with a standalone `go.mod` file. Consequentially, our normal
linting run won't process its sources as golangci-lint only supports
linting a single module at a time. To work around this, we thus have a
separate Makefile target that is supposed to run golangci-lint inside of
that separate module.

With Go 1.20, this linting target started to fail in our CI though with
the following error:

```
./packed_binaries.go:19:3: invalid go:embed: build system did not supply embed configuration
```

The error is confusing as the linter isn't even supposed to process the
reported file. But as it turns out, our linting Makefile target switches
directories and executes golangci-lint in separate shells, which causes
us to wrongly lint the main Gitaly module again. And given that this
Makefile target doesn't have our binaries as a dependency they might not
yet have been built, which then causes the `go:embed` line to complain.

The reason this started to fail is an upstream regression that was
introduced into Go via c0ac33c006 (cmd/go: don't compute Embed fields if
they're not needed, 2023-02-20). With this change, Go doesn't seem to
compute the `Embed` fields at all anymore, and thus golangci-lint starts
to complain.

Fix this bug by changing the directory in the same shell as we run the
linter command.