limiter: Add adaptive limit to the concurrency limiter
This commit adds the adaptive limit to limiter.ConcurrencyLimiter. It
changes the signature of the initializer to receive an AdaptiveLimit
object and converts the queues in the limiter to dynamic pools. At this
point, as the calculator has not been integrated (yet), the input limit
never changes. It always returns a persistent limit, which is a perfect
replacement for the integer static limit.

In the future, when the calculator kicks in, the dynamic pools in the
concurrency limiter scale up and down accordingly when the current limit
is changed by the calculator.