limiter: Generalize concurrency limiter's buffers used for semaphore
The concurrency limiter uses two buffered channels to control the
concurrency and queueing mechanims. Naturally, we can consider them as
two weighted semaphore. This commit extracts the core operations of
those two semaphores to a staticSemaphore struct. This extraction also
standardize the interface so that the forward commits can swap in
another semaphore implementation.