catfile: Introduce a state-machine based commit parser
Currently we have multiple sources for parsing commit messages:
1. `catfile` package has `parser` which implements `ParseCommits`
2. `commit` package has `extractSignature`

While the end usage of the two implementations may vary, they implement
logic to parse a given commit message. Parsing commit messages while
simple is ugly, because any errors can lead to vulnerabilities. Combined
with the fact that commits are more human readable than machine
readable, we need to ensure that our implementation is as close to git's
own while maintaining readability.

In this accord, implement a new parser in the `catfile` package, which
borrows logic from the existing parser but uses a simple state machine
to navigate and parse the commit.

This new implementation also fixes a security bug wherein we were
considering and parsing headers after the signature in the commit, while
not considering the headers in the payload of the commit.

This meant that malicious users could introduce headers into new
malformed commits while copying the rest of the data from existing
signed commits. The new commit would still be shown as signed by Gitaly
since we were simply skipping the malformed data but the headers would
be considered.

In the following commits, we will also remove the other parsers.