commit: Wait for the git-log(1) process to exit in FindCommits
In this commit, we add a new wait logic to wait for git-log(1)
command to finish;

Previously we did not Wait() for the git-log(1) process to exit
in FindCommits, causing us to fail to include its execution time
in the command details
and not log any errors it returned.

We differentiate benign errors from real errors.
Benign errors are caused by terminating the stream
since response limit is reached;
while real errors are caused by git log command failures.
Non-benign errors are logged as errors.

Changelog: fixed