repository: Make ApplyGitattributes RPC no-op in git 2.43
In git 2.43.0, we will be writing to`HEAD:.gitattributes` instead of
`info/gitattributes`, so ApplyGitattributes is no longer needed.
In this commit, we make ApplyGitattributes RPC no-op.

Beside that, ApplyGitattributes will also delete `info/gitattributes`.
This is because `info/gitattributes` has a higher precedence
than HEAD:.gitattributes. We want to avoid `HEAD:.gitattributes` being
overridden.