backup: Add ResetRefs to the Repository interface
Modifies the Repository interface to add a function called `ResetRefs`
which resets the references in the repository to mirror a provided list
of references.

This will be used to optimise repository restores, as we can simply
apply the list of refs from the backup to the existing repo rather than
recreating the repo from the backed-up bundle.

Additionally, the `NewRemoteRepository` and `NewLocalRepository`
constructors have exported to facilitate unit testing. I dislike this
approach, but we get an import cycle in `repository_test.go` otherwise,
since it'll need to be defined within the `backup` package instead of
`backup_test`, and it imports the `testserver` package which in turn
imports the `backup` package.