backup: Fix dangling repo check
Addresses a few issues with the prior implementation that checked for
"dangling" repositories after a restore. A dangling repo is one that
was created after a backup which is being restored was made, and should
be deleted to keep Rails and Gitaly in sync.

Modifies the `processedRepos` inner map so it's keyed by a string
instead of a pointer value. Comparing repository objects via their
pointers would not work correctly. Also fixes how we're checking if a
key exists in the nested inner map.

Fixes the existing tests so that all repositories are created with
`gittest.CreateRepository` and are thus visible when we later query
`ListRepositories` in the restore logic.