Remove an obsolete TransactionManager test
TransactionManager has a test asserting that the pack file includes
unreachable objects. This was done previously as we didn't have a
good way to pack and hold on to all needed objects to ensure the packs
will always apply.

We've since introduced new functionality in Git that allows us to pack
the objects from the quarantine and figure out at the same time which
objects the quarantined objects depend on in the repository. We'll thus
prevent pruning the unreachable objects that a transaction depended upon
instead of packing them in the pack.

As we're about to change the behavior, remove the test that is about to
be obsoleted.