Generalize packRefsDirectoryEntry to anyDirectoryEntry
Previously, we introduce packRefsDirectoryEntry to assert the on-disk
content of pack-refs file in WAL's luggage. That assertion verify if
that file exists but don't look into its content. The content is verifed
in a later stage. We'll do the same for packfiles. This commit
generalize this assert helper for re-using in later commits.