Add an assertion to assert state of packfiles in transactions
This commit implements new asserter for a more detailed examining
objects of a repository. This assert helper allows the tests to:
- Assert all packfiles in objects/pack repository.
- Assert the content and verify indexes of each packfile.
- Assert multi-pack-index file.
- Assert "invisible" objects which don't locate on disk but present from
  Git's perspective. Typically, it's used in an alternate setting where
  pool member can see other objects of the pool.

This asserter is a superset of the existing Objects asserter. However,
it's too verbose. A majority of the tests should not be concerned with
the on-disk packfile layout of objects. So, the existing Objects
asserter is kept.