Add WalkObjects to walk the object graph
The TransactionManager needs to walk objects in the quarantine to
identify which objects to commit with the transaction in a packfile
and which objects that already exist in the repository the transaction
depends on.

This commit adds WalkObjects to facilitate that. It walks the object
graph from given start points and prints out all objects IDs that it
encounters. If applicable, the objects path is also printed out to aid
packing them later.

Objects that are encountered but do not exist are printed out with a '?'
to mark them missing. When the WalkObjects is invoked with just the
quarantine directory without the alternates, this leads to the object walk
identifying the objects from the quarantine that we need to pack with the
missing objects being the transaction's dependencies that must exist in
the repository already.