Simplify checking whether packing is needed
TransactionManager is currently looking into the quarantine directory
explicitly to see if there are new objects. As we're now walking the
new reference tips in the quarantine directory, we know whether we
need to pack objects or not based on whether we get object IDs back
from the walk. Let's thus simplify and remove the additional check.
As an optimization, we only launch `pack-objects` and `index-pack` if
we receive an object to pack. This avoids launching the commands for
transactions that don't need any new objects from the quarantine to
be committed.

As all missing objects are now noticed while verifying the object
dependencies of the transaction, the returned error changes in one
of the test cases. The error cause is still the same though.